[REF] html_editor: remove ZWS and InsertText plugins
This refactor has two main goals: group the responsibilities of handling
empty inlines in a single plugin and reduce the number of core plugins.

Before this commit, empty inline elements were handled by the Format and
ZWS plugins. The former added a ZWS (zero witdh space character) to
empty inlines and added a attribute to mark them as empty. The latter
was responsible for cleaning the ZWS characters and removing the
mentioned attribute once those inlines were no longer empty.

This commit groups such responsibilities in the Format plugin, allowing
us to remove the ZWS plugin.

The InsertText plugin is also removed, moving the responsability of
deleting the contents of a non-collapsed selection before input to the
Delete plugin.
