+++ addons/html_editor/static/src/main/inline_code.js
@@ -5,8 +5,9 @@ import { DIRECTIONS } from "@html_editor/utils/position";
 export class InlineCodePlugin extends Plugin {
     static name = "inline_code";
     static dependencies = ["selection", "split"];
+    /** @type { (p: InlineCodePlugin) => Record<string, any> } */
     static resources = (p) => ({
-        onInput: p.onInput.bind(p),
+        onInput: { handler: p.onInput.bind(p) },
     });
 
     onInput(ev) {
