+++ addons/mail/static/src/views/web/fields/html_composer_message_field/mention_plugin.js
@@ -22,7 +22,7 @@ export class MentionPlugin extends Plugin {
     static dependencies = ["overlay", "dom", "history"];
 
     static resources = (p) => ({
-        onBeforeInput: p.onBeforeInput.bind(p),
+        onBeforeInput: { handler: p.onBeforeInput.bind(p) },
     });
 
     setup() {
